test: separate recorder selection domain coverage
